+++ packages/api/src/derives/session/index.ts
@@ -1,0 +1,2 @@
+export {queryKeyInfo as keyInfo, DerivedSessionKeyInfo} from './keyInfo';
+export * from '@polkadot/api-derive/session';
