+++ packages/api/src/derives/session/keyInfo.ts
@@ -1,0 +1,58 @@
+import {ApiInterfaceRx} from '@cennznet/api/types';
+import {memo} from '@polkadot/api-derive/util';
+import {createType, Option, Vec} from '@polkadot/types';
+import {AccountId, Keys} from '@polkadot/types/interfaces';
+import {ITuple} from '@polkadot/types/types';
+import {combineLatest, Observable, of} from 'rxjs';
+import {switchMap} from 'rxjs/operators';
+
+export interface DerivedSessionKeyInfo {
+  nextSessionKeys: AccountId[];
+  sessionKeys: AccountId[];
+}
+
+function unwrapSessionIds(
+  stashId: AccountId,
+  queuedKeys: [AccountId, Keys][],
+  nextKeys: Option<Keys>
+): {nextSessionKeys: AccountId[]; sessionKeys: AccountId[]} {
+  let sessionKeys: AccountId[] = [];
+  const idKeys = queuedKeys.find(([currentId]): boolean => currentId.eq(stashId));
+  if (idKeys) {
+    sessionKeys = idKeys[1];
+  }
+  const nextSessionKeys = nextKeys.unwrapOr([] as AccountId[]);
+
+  return {
+    nextSessionKeys,
+    sessionKeys,
+  };
+}
+
+function retrieveSessionDetails(
+  api: ApiInterfaceRx,
+  stashId: AccountId
+): Observable<[Vec<ITuple<[AccountId, Keys]>>, Option<Keys>]> {
+  return combineLatest([
+    api.query.session.queuedKeys<Vec<ITuple<[AccountId, Keys]>>>(),
+    api.query.session.nextKeys(stashId),
+  ]) as Observable<[Vec<ITuple<[AccountId, Keys]>>, Option<Keys>]>;
+}
+
+/**
+ * @description From a stash and sessions nextKeys, filter session and next session details
+ */
+export function queryKeyInfo(api: ApiInterfaceRx): (accountId: Uint8Array) => Observable<DerivedSessionKeyInfo> {
+  return memo(
+    (accountId: Uint8Array | string): Observable<DerivedSessionKeyInfo> => {
+      const stashId = createType(api.registry, 'AccountId', accountId);
+      return retrieveSessionDetails(api, stashId).pipe(
+        switchMap(
+          ([queuedKeys, nextKeys]): Observable<DerivedSessionKeyInfo> => {
+            return of(unwrapSessionIds(stashId, queuedKeys, nextKeys));
+          }
+        )
+      );
+    }
+  );
+}
