+++ packages/api/src/derives/staking/index.ts
@@ -1,9 +1,4 @@
-export {
-  queryStakingAccountInfo as accountInfo,
-  querySessionInfo as sessionKeyInfo,
-  DerivedStakingInfo,
-  DerivedSessionKeyInfo,
-} from './stakingAccount';
+export {queryStakingAccountInfo as accountInfo, DerivedStakingInfo} from './stakingAccount';
 export * from '@polkadot/api-derive/staking/account';
 export * from '@polkadot/api-derive/staking/controllers';
 export * from '@polkadot/api-derive/staking/electedInfo';
